배너 관리/캐싱 기능 구현 완료

+++ src/main/java/com/example/lotteon/entity/admin/banner/BannerInfo.java
@@ -19,11 +19,11 @@ public class BannerInfo {
 
   @Field("width")
   @SerializedName("width")
-  private int width;
+  private double width;
 
   @Field("height")
   @SerializedName("height")
-  private int height;
+  private double height;
 
   @Field("background_color")
   @SerializedName("background_color")
